Мелкие исправления
Исправлено:
- Изменено название поля в классе LevelThresholds с sourse_id на sourceId
- Изменен тип данных у поля creditVolume с int на double в классе DailyResult
- В классе Task было добавлено поле targetValue в конструктор

+++ src/main/java/com/example/sber/model/entity/DailyResult.java
@@ -17,13 +17,12 @@ public class DailyResult {
 
     @ManyToOne
     private Employee employee;
-
     private int dealsCount;
-    private int creditVolume;
+    private double creditVolume;
     private int productCount;
     private LocalDate date;
 
-    public DailyResult (Long id, Employee employee, int dealsCount, int creditVolume, int productCount, LocalDate date) {
+    public DailyResult (Long id, Employee employee, int dealsCount, double creditVolume, int productCount, LocalDate date) {
         this.id = id;
         this.employee = employee;
         this.dealsCount = dealsCount;
